Use Long type

+++ src/main/java/com/mit/mahendracandi/atm/dto/LoginResponse.java
@@ -1,6 +1,4 @@
 package com.mit.mahendracandi.atm.dto;
 
-import java.math.BigDecimal;
-
-public record LoginResponse(String accountNumber, String name, BigDecimal balance) {
+public record LoginResponse(String accountNumber, String name, Long balance) {
 }
